+++ WebStore/Views/Employee/Index.cshtml
@@ -8,7 +8,7 @@
                 <td>@item.Surname</td>
                 <td>@item.Name</td>
                 <td>@item.Age</td>
-                <td><a href="~/Home/Employee/@item.Id">Подробнее</a></td>
+                <td><a asp-controller="Employee" asp-action="Employee" asp-route-id=@item.Id>Подробнее</a></td>
             </tr>
         }
     </table>
