--- WebStore/Views/Home/Employee.cshtml
@@ -1,14 +1,0 @@
-﻿@model WebStore.Models.Employee
-
-<!DOCTYPE html>
-
-<html>
-<head>
-    <meta name="viewport" content="width=device-width" />
-    <title>Сотрудник №@Model.Id</title>
-</head>
-<body>
-    <p>@Model.ToString()</p>
-    <a href="~/Home/Employees">Назад</a>
-</body>
-</html>
