+++ WebStore/Views/Shared/Partial/_MainMenu.cshtml
@@ -20,6 +20,6 @@
         </li>
         <li><a asp-controller="Home" asp-action="Error404">404</a></li>
         <li><a asp-controller="Home" asp-action="ContactUs">Contact</a></li>
-        <li><a asp-controller="Home" asp-action="Employees">Employees</a></li>
+        <li><a asp-controller="Employee" asp-action="Index">Employees</a></li>
     </ul>
 </div>
